enable higher-order extension functions
